+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Introduce async/await for async requests. Use axios instead of j-fetch
+([#100](https://github.com/fs/react-base/pull/100))
 - Add "env" preset and move some plugins to production environment
 ([#98](https://github.com/fs/react-base/pull/98))
 - Setup HMR for reducers